Refactor providers to single file and use them in Address-, Event-, Service-, and UnitItem tests
