+++ src/components/ListItems/ServiceItem/__tests__/ServiceItem.test.js
@@ -1,12 +1,7 @@
 import React from 'react';
-import { ThemeProvider } from '@mui/material/styles';
-import { IntlProvider } from 'react-intl';
-import { Provider } from 'react-redux';
-import configureStore from 'redux-mock-store';
-import { render } from '@testing-library/react';
-import themes from '../../../../themes';
 import { initialState } from '../../../../redux/reducers/user';
 import ServiceItem from '../index';
+import { getRenderWithProviders } from '../../../../../jestUtils';
 
 const mockData = {
   "name": {
@@ -23,42 +18,13 @@ const mockProps = {
   service: mockData,
 };
 
-// Mock props for intl provider
-const intlMock = {
-  locale: 'en',
-  messages: {
-    'unit.accessibility.noInfo': 'No info',
-    'unit.accessibility.ok': 'Accessibility ok',
-    'unit.accessibility.problems': '{count} accessibility problems'
+const renderWithProviders = getRenderWithProviders({
+  user: initialState,
+  settings: {},
+  service: {
+    current: null,
   },
-};
-
-// simpleItem={embeddedList} key={`unit-${id}`} className={`unit-${id}`} unit={item}
-
-const mockStore = configureStore([]);
-
-// eslint-disable-next-line react/prop-types
-const Providers = ({ children }) => {
-  const store = mockStore({
-    user: initialState,
-    settings: {},
-    service: {
-      current: null,
-    },
-  });
-
-  return (
-    <Provider store={store}>
-      <IntlProvider {...intlMock}>
-        <ThemeProvider theme={themes.SMTheme}>
-          {children}
-        </ThemeProvider>
-      </IntlProvider>
-    </Provider>
-  );
-};
-
-const renderWithProviders = component => render(component, { wrapper: Providers });
+})
 
 describe('<ServiceItem />', () => {
   it('should work', () => {
